+++ broker/src/integration-test/groovy/com/swisscom/cloud/sb/broker/backup/shield/ShieldBackupRestoreProviderSpec.groovy
@@ -94,7 +94,7 @@ class ShieldBackupRestoreProviderSpec extends BaseTransactionalSpecification {
         given:
         String taskId = 'task-uuid'
         Backup backup = new Backup(serviceInstanceGuid: 'service-guid', operation: Backup.Operation.CREATE, externalId: taskId)
-        shieldClient.getJobStatus(backup.externalId) >> JobStatus.SUCCESSFUL
+        shieldClient.getJobStatus(backup.externalId, backup) >> JobStatus.SUCCESSFUL
         when:
         Backup.Status status = shieldBackupRestoreProvider.getBackupStatus(backup)
         then:
@@ -105,7 +105,7 @@ class ShieldBackupRestoreProviderSpec extends BaseTransactionalSpecification {
         given:
         String taskId = 'task-uuid'
         Backup backup = new Backup(serviceInstanceGuid: 'service-guid', operation: Backup.Operation.CREATE, externalId: taskId)
-        shieldClient.getJobStatus(backup.externalId) >> { throw new ShieldResourceNotFoundException("Doesntmatter") }
+        shieldClient.getJobStatus(backup.externalId, backup) >> { throw new ShieldResourceNotFoundException("Doesntmatter") }
         when:
         Backup.Status status = shieldBackupRestoreProvider.getBackupStatus(backup)
         then: