+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/binding/FetchServiceBindingProvider.groovy
@@ -1,0 +1,7 @@
+package com.swisscom.cloud.sb.broker.binding
+
+import com.swisscom.cloud.sb.broker.model.ServiceBinding
+
+interface FetchServiceBindingProvider {
+    ServiceInstanceBindingResponseDto fetchServiceBinding(ServiceBinding serviceBinding)
+}