+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 
+## [2.12.0] - 2018-04-13
+- Fixed issue that not all flags were correctly returned in the catalog (*bindable)
+- Support custom service provider service binding fetch
+- Support policy based random string generation
+- Support retrying backup in case of failure
+
 ## [2.11.6] - 2018-04-11
 - Add dto validation helper class for serviceproviders
 - Support Openshift deployment